pyproject.toml
[project]
name = "nillabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.64.0",
    "pytelegrambotapi>=4.31.0",
    "python-dotenv>=1.2.1",
]
